Project1 Final Version

+++ CarRentalApp/src/main/java/com/example/CarRentalApp/repository/LocationRepo.java
@@ -3,5 +3,5 @@ package com.example.CarRentalApp.repository;
 import com.example.CarRentalApp.model.Location;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocationRepo extends JpaRepository<Location, Long> {
+public interface LocationRepo extends JpaRepository<Location, Integer> {
 }
